Move unit tests to own project
